:recycle: Refactoring for PickPage rendering optimization #2
