+++ src/pages/BestCombinationPickPage.tsx
@@ -1,33 +1,15 @@
-import useToggleFilter from '@hooks/useToggleFilter';
 import DeskTopImage from '@components/BestCombinationPick/DeskTopImage';
-import AttributeButtonListContainer from '@components/BestCombinationPick/AttributeButtonListContainer';
-import NavigateListPageButton from '@components/BestCombinationPick/NavigateListPageButton';
+import AttributePick from '@components/BestCombinationPick/AttributePick';
 import Wrapper from '@components/Common/UI/Wrapper';
-import { 꿀조합_픽_초기_필터 } from '@constants/constants';
 import styled from '@emotion/styled';
 import { flexbox } from '@styles/mixin';
 import mediaQuery from '@styles/media-queries';
 
 function BestCombinationPickPage() {
-  const {
-    selectedFilter: 선택된_꿀조합_속성,
-    overSelectedFilter,
-    toggleFilter: 클릭핸들러_꿀조합_속성_토글,
-    initializeFilter: 클릭핸들러_꿀조합_속성_초기화,
-  } = useToggleFilter(꿀조합_픽_초기_필터);
-
   return (
     <TotalContainer>
       <DeskTopImage />
-      <Container>
-        <AttributeButtonListContainer
-          filteredAttr={선택된_꿀조합_속성}
-          overSelectedFilter={overSelectedFilter}
-          toggleFilter={클릭핸들러_꿀조합_속성_토글}
-          initializeFilter={클릭핸들러_꿀조합_속성_초기화}
-        />
-        <NavigateListPageButton filteredAttr={선택된_꿀조합_속성} />
-      </Container>
+      <AttributePick />
     </TotalContainer>
   );
 }
@@ -39,9 +21,4 @@ const TotalContainer = styled(Wrapper)`
     min-height: 750px;
     position: relative;
   }
-`;
-const Container = styled(Wrapper)`
-  ${flexbox('column', 'space-between', 'center')};
-  max-width: 600px;
-  padding: 26px;
 `;
